+++ Src/MoneyFox.Business/ViewModels/ModifyPaymentViewModel.cs
@@ -127,7 +127,7 @@ namespace MoneyFox.Business.ViewModels
         {
             SelectedPayment = new PaymentViewModel
             {
-                Type = (int) paymentType,
+                Type = paymentType,
                 Date = DateTime.Now,
                 // Assign empty CategoryViewModel to reset the GUI
                 Category = new CategoryViewModel(),
@@ -445,7 +445,7 @@ namespace MoneyFox.Business.ViewModels
         ///     Returns the Header for the AccountViewModel field
         /// </summary>
         public string AccountHeader
-            => SelectedPayment?.Type == (int) PaymentType.Income
+            => SelectedPayment?.Type == PaymentType.Income
                 ? Strings.TargetAccountLabel
                 : Strings.ChargedAccountLabel;
